LayoutTests/http/wpt/html/browsers/windows/browsing-context-expected.txt


FAIL Check that browsing context has new, ready HTML document assert_equals: The document's encoding should be 'UTF-8'. expected "UTF-8" but got "windows-1252"
PASS Check that new document nodes extant, empty
PASS Check the document properties corresponding to the creator browsing context
